+++ src/FluentCertificates.Common/Oids.cs
@@ -250,6 +250,8 @@ public static class Oids
     public const string ApplicationCertPolicies = "1.3.6.1.4.1.311.21.10";
     /// <summary>Cryptographic attribute type OID <c>1.3.6.1.5.5.7.1.1</c>.</summary>
     public const string AuthorityInformationAccess = "1.3.6.1.5.5.7.1.1";
+    /// <summary>Cryptographic attribute type OID <c>1.3.6.1.5.5.7.1.11</c>.</summary>
+    public const string SubjectInformationAccess = "1.3.6.1.5.5.7.1.11";
     /// <summary>Cryptographic attribute type OID <c>1.3.6.1.5.5.7.48.1</c>.</summary>
     public const string OcspEndpoint = "1.3.6.1.5.5.7.48.1";
     /// <summary>Cryptographic attribute type OID <c>1.3.6.1.5.5.7.48.2</c>.</summary>
@@ -471,6 +473,8 @@ public static class Oids
     public const string CertPolicyConstraints = "2.5.29.36";
     /// <summary>Certificate extension OID <c>2.5.29.37</c>.</summary>
     public const string EnhancedKeyUsage = "2.5.29.37";
+    /// <summary>Certificate extension OID <c>2.5.29.46</c>.</summary>
+    public const string FreshestCrl = "2.5.29.46";
     /// <summary>Certificate extension OID <c>2.5.29.54</c>.</summary>
     public const string InhibitAnyPolicyExtension = "2.5.29.54";
 
@@ -774,6 +778,7 @@ public static class Oids
     private static Oid? s_certificateTemplateOid;
     private static Oid? s_applicationCertPoliciesOid;
     private static Oid? s_authorityInformationAccessOid;
+    private static Oid? s_subjectInformationAccessOid;
     private static Oid? s_ocspEndpointOid;
     private static Oid? s_certificateAuthorityIssuersOid;
     private static Oid? s_pkcs9ExtensionRequestOid;
@@ -885,6 +890,7 @@ public static class Oids
     private static Oid? s_authorityKeyIdentifierOid;
     private static Oid? s_certPolicyConstraintsOid;
     private static Oid? s_enhancedKeyUsageOid;
+    private static Oid? s_freshestCrlOid;
     private static Oid? s_inhibitAnyPolicyExtensionOid;
 
     private static Oid? s_tstInfoOid;
@@ -1185,6 +1191,8 @@ public static class Oids
     public static Oid ApplicationCertPoliciesOid => LazyInitializer.EnsureInitialized(ref s_applicationCertPoliciesOid, () => InitializeOid(ApplicationCertPolicies));
     /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="AuthorityInformationAccess"/>.</summary>
     public static Oid AuthorityInformationAccessOid => LazyInitializer.EnsureInitialized(ref s_authorityInformationAccessOid, () => InitializeOid(AuthorityInformationAccess));
+    /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="SubjectInformationAccess"/>.</summary>
+    public static Oid SubjectInformationAccessOid => LazyInitializer.EnsureInitialized(ref s_subjectInformationAccessOid, () => InitializeOid(SubjectInformationAccess));
     /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="OcspEndpoint"/>.</summary>
     public static Oid OcspEndpointOid => LazyInitializer.EnsureInitialized(ref s_ocspEndpointOid, () => InitializeOid(OcspEndpoint));
     /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="CertificateAuthorityIssuers"/>.</summary>
@@ -1398,6 +1406,8 @@ public static class Oids
     public static Oid CertPolicyConstraintsOid => LazyInitializer.EnsureInitialized(ref s_certPolicyConstraintsOid, () => InitializeOid(CertPolicyConstraints));
     /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="EnhancedKeyUsage"/>.</summary>
     public static Oid EnhancedKeyUsageOid => LazyInitializer.EnsureInitialized(ref s_enhancedKeyUsageOid, () => InitializeOid(EnhancedKeyUsage));
+    /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="FreshestCrl"/>.</summary>
+    public static Oid FreshestCrlOid => LazyInitializer.EnsureInitialized(ref s_freshestCrlOid, () => InitializeOid(FreshestCrl));
     /// <summary>A shared, cached <see cref="Oid"/> instance for <see cref="InhibitAnyPolicyExtension"/>.</summary>
     public static Oid InhibitAnyPolicyExtensionOid => LazyInitializer.EnsureInitialized(ref s_inhibitAnyPolicyExtensionOid, () => InitializeOid(InhibitAnyPolicyExtension));